Swap TLS signatures in case we are signing them with GOST keys
GOST TLS suites have one peculiarity: it is the only place, where the
endianness of the signature is switched. Instead of s-LE, r-LE, here
standards demand using r-BE, s-BE. Again. This is the only place, where
such change is required. So, instead of adding additional code to pk.c,
just swap bytes in signatures in TLS code.

Signed-off-by: Dmitry Eremin-Solenikov <dbaryshkov@gmail.com>